Move all code files into src folder
