+++ src/components/camera/CameraScreen.tsx
