+++ src/components/camera/PhotoActionButtons.tsx
