+++ src/components/camera/PhotoPreview.tsx
