+++ src/components/events/EventDetails.tsx
